[Silver III] Title: 낚이고 낚아라, Time: 384 ms, Memory: 3688 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/15821. 낚이고 낚아라/낚이고 낚아라.cc
@@ -29,9 +29,9 @@ int main() {
         cin >> a;
         ll mx = 0;
         for (int j = 0; j < a; j++) {
-            int x, y;
+            ll x, y;
             cin >> x >> y;
-            mx = max(mx, (ll)(x*x + y*y));
+            mx = max(mx, (x*x + y*y));
         }
         v.push_back(mx);
     }
